Split config_structs into 2 separate objects
