+++ src/gbm_exception.h
